Initialize the project with a data converter
MDTCal is the next version of Sigcal. We are re-naming the project to
reflect the big changes to the structure of the code and more
importantly, the calibration process work flow.  This initialization
features a revised data converter for Symmetric Research data files that
bypasses Symres's sequential file processing, and it creates a separate
SAC file for each .DAT file and channel. This will eliminate the need to
hand-edit the folder after making the sin sweep, & remove every other
file in order to separate them into stationary files.