Add the masters thesis to the project for reference.
Masters thesis for the technique of mass displacement tracking and how it is used to calibrate an electromechanical seismometer. This software is the tool to accompany the seismometer/digitizer along with a laser position sensor in order to calibrate the electromechanical seismometer and output a series of poles and zeros that describe the seismometer response. With the software, and the elicited response, a dataless SEED file for the station channel can be generated.